Added sync finished checks and single task/task list functions
GTSyncManager isSyncing property now gets set to NO when sync completes.

Added public methods to modify a single task or task list on the
server.  The change will only be performed if a sync is not currently
in process.  Each method returns YES if the operation was attempted,
otherwise NO.